+++ crates/moon/tests/test_cases/moon_new/plain/lib/hello.mbt
