+++ crates/moon/tests/test_cases/moon_new/plain/lib/moon.pkg.json
