+++ crates/moon/tests/test_cases/moon_new/plain/main/moon.pkg.json
