+++ crates/moon/tests/test_cases/moon_new/plain/moon.mod.json
